+++ src/modules/search/components/index.ts
@@ -1,1 +1,4 @@
 export { CommandPalette } from './CommandPalette';
+export { SearchPanel } from './SearchPanel';
+export { SearchResults } from './SearchResults';
+export { FacetSidebar } from './FacetSidebar';