+++ src/modules/search/index.ts
@@ -1,5 +1,5 @@
 // Components
-export { CommandPalette } from './components';
+export { CommandPalette, SearchPanel, SearchResults, FacetSidebar } from './components';
 
 // Hooks
 export {
@@ -8,7 +8,22 @@ export {
   type CommandPaletteResult,
   useKeywordSearch,
   type UseKeywordSearchReturn,
+  useSemanticSearch,
+  type UseSemanticSearchReturn,
+  useHybridSearch,
+  type UseHybridSearchReturn,
+  useSearchPanel,
+  type UseSearchPanelReturn,
 } from './hooks';
 
 // Services
-export { KeywordSearchService } from './services';
+export { KeywordSearchService, SemanticSearchService, HybridSearchService, FacetService } from './services';
+
+// Types
+export {
+  type Facet,
+  type FacetValue,
+  type FacetSelection,
+  type DateRangeBucket,
+  DATE_BUCKET_LABELS,
+} from './types';