+++ src/modules/search/services/index.ts
@@ -1,3 +1,4 @@
 export { KeywordSearchService } from './KeywordSearchService';
 export { SemanticSearchService } from './SemanticSearchService';
 export { HybridSearchService, applyRRF } from './HybridSearchService';
+export { FacetService } from './FacetService';